Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -47,7 +47,7 @@ def get_average_age_for_season(data, season)
   counter = 0
   ageTotal = 0
   data[season].each {|elemObj|
-    ageTotal += elemObj["age"].to_f
+    ageTotal += elemObj["age"].to_i
     counter += 1
   }
   ageTotal/counter
